Add temppath as parameter
